Show health bars only for active players thanks to new script
