+++ Assets/Scripts/UI/UIHealthBar.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: f3ffa1a9ce7ec427caa183bac8275f2f
-timeCreated: 1493666715
+guid: 2a5fde9c2354c344ca595698588bfc39
+timeCreated: 1496503715
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
